+++ open-metadata-resources/open-metadata-samples/open-metadata-security-samples/pom.xml
@@ -27,6 +27,21 @@
         Samples showing how to configure an OMAG Service Platform for security.
     </description>
 
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-site-plugin</artifactId>
+                <version>3.9.1</version>
+            </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-project-info-reports-plugin</artifactId>
+                <version>3.1.1</version>
+            </plugin>
+        </plugins>
+    </build>
+
     <dependencies>
 
         <dependency>
