+++ open-metadata-resources/open-metadata-samples/open-metadata-security-samples/src/main/java/org/odpi/openmetadata/metadatasecurity/samples/CocoPharmaServerSecurityProvider.java
@@ -18,7 +18,7 @@ public class CocoPharmaServerSecurityProvider extends OpenMetadataServerSecurity
     {
         super();
 
-        Class<?>        connectorClass = CocoPharmaServerSecurityConnector.class;
+        Class<?>        connectorClass = AbstractCocoPharmaServerSecurityConnector.class;
 
         super.setConnectorClassName(connectorClass.getName());
     }
